+++ packages/expo-router/src/global-state/routing.ts
@@ -196,6 +196,7 @@ function getNavigateAction(
   const rootPayload: Record<string, any> = { params: {} };
   let payload = rootPayload;
   let params = payload.params;
+  let { screen, ...paramsWithoutScreen } = params;
 
   // The root level of payload is a bit weird, its params are in the child object
   while (actionStateRoute) {
@@ -204,9 +205,10 @@ function getNavigateAction(
 
     actionStateRoute = actionStateRoute.state?.routes[actionStateRoute.state?.routes.length - 1];
 
-    payload.params ??= {};
+    payload.params = Object.assign({}, { ...paramsWithoutScreen }, payload.params);
     payload = payload.params;
     params = payload;
+    ({ screen, ...paramsWithoutScreen } = params);
   }
 
   // Expo Router uses only three actions, but these don't directly translate to all navigator actions